requirements.txt
google-generativeai===0.5.0
twilio===9.0.4
flask===3.0.3